requirements.txt
youtube_transcript_api
streamlit
google_generativeai
python_dotenv
pathlib
fpdf